+++ benchmark/CMakeLists.txt
@@ -2,7 +2,6 @@ cmake_minimum_required(VERSION 3.14)
 project(phtree-benchmarks)
 
 set(BENCHMARK_ENABLE_TESTING OFF)
-#set(BENCHMARK_ENABLE_INSTALL ON)
 
 include(FetchContent)
 
@@ -30,9 +29,7 @@ macro(package_add_benchmark TESTNAME)
     target_include_directories(${TESTNAME} PRIVATE ${PROJECT_SOURCE_DIR}/..)
 endmacro()
 
-if (MSVC)
-    add_compile_definitions(RUN_HAVE_STD_REGEX=0 RUN_HAVE_POSIX_REGEX=0 COMPILE_HAVE_GNU_POSIX_REGEX=0)
-endif ()
+add_compile_definitions(RUN_HAVE_STD_REGEX=0 RUN_HAVE_POSIX_REGEX=0 COMPILE_HAVE_GNU_POSIX_REGEX=0)
 
 package_add_benchmark(count_mm_d_benchmark count_mm_d_benchmark.cc)
 package_add_benchmark(erase_benchmark erase_benchmark.cc)
